Add README.md, move db initialisation to mock directory
